+++ android/src/main/java/ly/count/dart/countly_flutter/CountlyFlutterPlugin.java
@@ -1644,6 +1644,10 @@ public class CountlyFlutterPlugin implements MethodCallHandler, FlutterPlugin, A
             this.config.setSDKBehaviorSettings(_config.getString("sdkBehaviorSettings"));
         }
 
+        if (_config.has("sdkBehaviorSettingsUpdatesDisabled")) {
+            this.config.disableSDKBehaviorSettingsUpdates();
+        }
+
         // APM ------------------------------------------------
         if (_config.has("trackAppStartTime")) {
             this.config.apm.enableAppStartTimeTracking();